actualizacion

+++ app.py
@@ -108,7 +108,7 @@ def recommendations(title):
     release_year = df_f.loc[title, 'release_year']
     df_f = df_f[df_f['release_year'].between(release_year - 2, release_year + 2)]
 
-    count = CountVectorizer(dtype=np.int16)
+    count = CountVectorizer(dtype=np.int8)
     count_matrix = count.fit_transform(df_f['soup'])
     indexes = pd.Series(df_f.index)
